[Silver II] Title: 통계학, Time: 72 ms, Memory: 4004 KB -BaekjoonHub

+++ 백준/Silver/2108. 통계학/통계학.cc
@@ -25,7 +25,7 @@ int main() {
 
     sort(nums, nums + N);
 
-    int avg = (int)round(sum / N);
+    int avg = round(sum / N);
     int mid = nums[N / 2];
 
     int mx = 0, idx1 = -1, idx2 = -1;
